requirements.txt
apiclient==1.0.4
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
exceptiongroup==1.2.2
filelock==3.18.0
fsspec==2025.7.0
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.1.1
oauth2client==4.1.3
oauthlib==3.2.2
openai-whisper==20250625
outcome==1.3.0.post0
pillow==10.4.0
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyparsing==3.1.4
PySocks==1.7.1
python-dotenv==1.0.1
regex==2025.7.34
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.25.0
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.14.0
tiktoken==0.11.0
torch==2.8.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
